--- internal/client_manager/client_manager.go
@@ -1,1 +1,0 @@
-package client_manager
