+++ internal/manager/models/client_db_model_test.go
@@ -1,4 +1,4 @@
-package client_manager
+package models
 
 import (
 	"testing"
